Define `NamedMatcherMetadata` as inner class of `NamedMatcher`
